benchmarks/requirements.txt
build==0.10.0
causal-learn==0.1.3.3
causalAssembly==0.5
certifi==2023.5.7
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
distinctipy==1.2.2
distlib==0.3.6
exceptiongroup==1.1.1
factor-analyzer==0.4.1
filelock==3.12.0
fonttools==4.39.4
future==0.18.3
gcastle==1.0.3
ghp-import==2.1.0
graphviz==0.20.1
griffe==0.29.0
identify==2.5.24
idna==3.4
igraph==0.10.4
iniconfig==2.0.0
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
lingam==1.7.1
lit==16.0.5.post0
Markdown==3.3.7
MarkupSafe==2.1.3
matplotlib==3.7.1
mergedeep==1.3.4
mike==1.1.2
mkdocs==1.4.3
mkdocs-autorefs==0.4.1
mkdocs-material==9.1.15
mkdocs-material-extensions==1.1.1
mkdocstrings==0.22.0
mkdocstrings-python==1.1.2
mpmath==1.3.0
networkx==3.1
nodeenv==1.8.0
numpy==1.23.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
packaging==23.1
pandas==2.0.2
patsy==0.5.3
Pillow==9.5.0
pip-tools==6.13.0
platformdirs==3.5.1
pluggy==1.0.0
pre-commit==3.3.2
progressbar2==4.2.0
pycparser==2.21
pydantic==1.10.8
pydot==1.4.2
pygam==0.8.0
Pygments==2.15.1
pymdown-extensions==10.0.1
pyparsing==3.0.9
pyproject_hooks==1.0.0
pytest==7.3.1
python-dateutil==2.8.2
python-utils==3.6.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyyaml_env_tag==0.1
regex==2023.6.3
requests==2.31.0
rpy2==3.5.12
ruff==0.0.270
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
statsmodels==0.14.0
sympy==1.12
texttable==1.6.7
threadpoolctl==3.1.0
tomli==2.0.1
torch==2.0.1
tqdm==4.65.0
triton==2.0.0
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==4.3
urllib3==2.0.2
verspec==0.1.0
virtualenv==20.23.0
watchdog==3.0.0
